requirements.txt
streamlit==1.38.0
supabase==2.6.0
python-dotenv==1.0.1

